refactor: move tailable cursor core test into misc cursor integration
